Separated different libsql versions into different import paths, adapted tests, added `@libsql/client-wasm` peer and dev dep
