+++ drizzle-orm/package.json
@@ -47,6 +47,7 @@
 		"@cloudflare/workers-types": ">=3",
 		"@electric-sql/pglite": ">=0.1.1",
 		"@libsql/client": ">=0.10.0",
+		"@libsql/client-wasm": ">=0.10.0",
 		"@neondatabase/serverless": ">=0.1",
 		"@op-engineering/op-sqlite": ">=2",
 		"@opentelemetry/api": "^1.4.1",
@@ -162,6 +163,7 @@
 		"@cloudflare/workers-types": "^4.20230904.0",
 		"@electric-sql/pglite": "^0.2.12",
 		"@libsql/client": "^0.10.0",
+		"@libsql/client-wasm": "^0.10.0",
 		"@miniflare/d1": "^2.14.4",
 		"@neondatabase/serverless": "^0.9.0",
 		"@op-engineering/op-sqlite": "^2.0.16",
