+++ drizzle-orm/src/libsql/driver.ts
@@ -1,4 +1,4 @@
-import { type Client, type Config, createClient, type ResultSet } from '@libsql/client';
+import type { Client, ResultSet } from '@libsql/client';
 import type { BatchItem, BatchResponse } from '~/batch.ts';
 import { entityKind } from '~/entity.ts';
 import { DefaultLogger } from '~/logger.ts';
@@ -11,7 +11,7 @@ import {
 } from '~/relations.ts';
 import { BaseSQLiteDatabase } from '~/sqlite-core/db.ts';
 import { SQLiteAsyncDialect } from '~/sqlite-core/dialect.ts';
-import { type DrizzleConfig, type IfNotImported, type ImportTypeError, isConfig } from '~/utils.ts';
+import type { DrizzleConfig } from '~/utils.ts';
 import { LibSQLSession } from './session.ts';
 
 export class LibSQLDatabase<
@@ -29,7 +29,7 @@ export class LibSQLDatabase<
 	}
 }
 
-function construct<
+export function drizzle<
 	TSchema extends Record<string, unknown> = Record<string, never>,
 >(client: Client, config: DrizzleConfig<TSchema> = {}): LibSQLDatabase<TSchema> & {
 	$client: Client;
@@ -62,61 +62,12 @@ function construct<
 	return db as any;
 }
 
-export function drizzle<
-	TSchema extends Record<string, unknown> = Record<string, never>,
-	TClient extends Client = Client,
->(
-	...params: IfNotImported<
-		Client,
-		[ImportTypeError<'@libsql/client'>],
-		[
-			TClient | string,
-		] | [
-			TClient | string,
-			DrizzleConfig<TSchema>,
-		] | [
-			(
-				& DrizzleConfig<TSchema>
-				& ({
-					connection: string | Config;
-				} | {
-					client: TClient;
-				})
-			),
-		]
-	>
-): LibSQLDatabase<TSchema> & {
-	$client: TClient;
-} {
-	if (typeof params[0] === 'string') {
-		const instance = createClient({
-			url: params[0],
-		});
-
-		return construct(instance, params[1]) as any;
-	}
-
-	if (isConfig(params[0])) {
-		const { connection, client, ...drizzleConfig } = params[0] as
-			& { connection?: Config; client?: TClient }
-			& DrizzleConfig<TSchema>;
-
-		if (client) return construct(client, drizzleConfig) as any;
-
-		const instance = typeof connection === 'string' ? createClient({ url: connection }) : createClient(connection!);
-
-		return construct(instance, drizzleConfig) as any;
-	}
-
-	return construct(params[0] as TClient, params[1] as DrizzleConfig<TSchema> | undefined) as any;
-}
-
 export namespace drizzle {
 	export function mock<TSchema extends Record<string, unknown> = Record<string, never>>(
 		config?: DrizzleConfig<TSchema>,
 	): LibSQLDatabase<TSchema> & {
 		$client: '$client is not available on drizzle.mock()';
 	} {
-		return construct({} as any, config) as any;
+		return drizzle({} as any, config) as any;
 	}
 }
